core/validatorapi: relax `SlotFromTimestamp` checks
Export `SlotFromTimestamp` instead of keeping two equal implementation.

Relax `SlotFromTimestamp` timestamp checks: if a timestamp comes before genesis time as fetched from the beacon node, fall back to `time.Now()` as timestamp.

This function is only ever used to check for validator registration validity, which has never been formalized and can be relaxed.
